+++ contracts/libraries/StructureData.sol
@@ -60,7 +60,7 @@ library StructureData {
         uint8 environment;
         StructureData.OptionState onGoing;
         StructureData.OptionState expired; 
-        mapping(uint16 => uint128) depositPriceAfterExpiryPerRound; 
+        mapping(uint16 => uint256) depositPriceAfterExpiryPerRound; 
         mapping(uint16 => bool) expiryLevelSkipped;
         mapping(address=>StructureData.UserState) userStates;
     }
